added google signin

+++ shoppie-client/src/components/UserLogin/Login.tsx
@@ -1,51 +1,89 @@
 "use client"
 
-import React, { useState } from 'react';
+import React from 'react';
 import { useRouter } from 'next/navigation';
 import { userService } from '@/src/services/apiUrls';
 import { useShoppieStore } from '@/src/store/shoppieStore';
+import Swal from 'sweetalert2';
+import Link from 'next/link';
+import {useForm} from 'react-hook-form'
+import GoogleButton from 'react-google-button';
+import useGoogleSignIn from '../GoogleSignIn';
+
+import * as Yup from 'yup';
+
+import { yupResolver } from '@hookform/resolvers/yup';
+
+
+const Schema=Yup.object().shape({
+  email:Yup.string().email('Invalid email').required('Email is required'),
+  password:Yup.string().required('Password is required')
+})
 
 interface FormData {
   email: string;
   password: string;
 }
 
+
+
 const Login: React.FC = () => {
   const router = useRouter();
 
+  const {googleData} = useGoogleSignIn()
+
+  const { register, handleSubmit,reset, formState: { errors, isValid } } = useForm<FormData>({
+    resolver: yupResolver(Schema),
+    mode: 'onChange',
+  });
+
   const {login }= useShoppieStore()
+
+  
 
 
   
 
-  const [formData, setFormData] = useState<FormData>({
-    email: '',
-    password: '',
-  });
+  
 
-  const handleInput = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setFormData({ ...formData, [name]: value });
-  };
 
-  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
+
+  const LoginForm = async (data:FormData) => {
+   
     try {
-      const res = await userService.login(formData);
+      const res = await userService.login(data);
+      
       if (res.status === 200) {
-        router.push('/');
+       Swal.fire({
+          position: 'center',
+          icon: 'success',
+          title: 'Login Successful',
+          showConfirmButton: false,
+          timer: 1500
+        })
+        router.push('/')
+        
         login();
         const tokens = JSON.stringify(res.data);
         console.log(res.data);
         localStorage.setItem('token', tokens);
       }
     } catch (err) {
-      console.error(err);
+      Swal.fire({
+        position: 'center',
+        icon: 'error',
+        title: 'Invalid credentials',
+        showConfirmButton: false,
+        timer: 1500
+      })
+    }finally{
+      reset()
     }
   };
 
   return (
     <div>
+      
       <section className="bg-gray-50 dark:bg-gray-900">
         <div className="flex flex-col items-center justify-center px-6 py-8 mx-auto md:h-screen lg:py-0">
           <a
@@ -64,7 +102,7 @@ const Login: React.FC = () => {
               <h1 className="text-xl font-bold leading-tight tracking-tight text-gray-900 md:text-2xl dark:text-white">
                 Sign in to your account
               </h1>
-              <form className="space-y-4 md:space-y-6" onSubmit={handleSubmit}>
+              <form className="space-y-4 md:space-y-6" onSubmit={handleSubmit(LoginForm)}>
                 <div>
                   <label
                     htmlFor="email"
@@ -74,14 +112,14 @@ const Login: React.FC = () => {
                   </label>
                   <input
                     type="email"
-                    name="email"
-                    id="email"
+                    
+                    
                     className="bg-gray-50 border border-gray-300 text-gray-900 rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
                     placeholder="name@company.com"
-                    required
-                    value={formData.email}
-                    onChange={handleInput}
+                  
+                    {...register('email')}
                   />
+                  {errors.email && <p className='text-red-500 text-sm'>{errors.email.message}</p>}
                 </div>
                 <div>
                   <label
@@ -92,35 +130,17 @@ const Login: React.FC = () => {
                   </label>
                   <input
                     type="password"
-                    name="password"
+                    
                     id="password"
-                    placeholder="••••••••"
+                    placeholder="•••••••"
                     className="bg-gray-50 border border-gray-300 text-gray-900 rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
-                    required
-                    value={formData.password}
-                    onChange={handleInput}
+                   
+                    {...register('password')}
                   />
+                  {errors.password && <p className='text-red-500 text-sm '>{errors.password.message}</p>}
                 </div>
                 <div className="flex items-center justify-between">
-                  <div className="flex items-start">
-                    <div className="flex items-center h-5">
-                      <input
-                        id="remember"
-                        aria-describedby="remember"
-                        type="checkbox"
-                        className="w-4 h-4 border border-gray-300 rounded bg-gray-50 focus:ring-3 focus:ring-primary-300 dark:bg-gray-700 dark:border-gray-600 dark:focus:ring-primary-600 dark:ring-offset-gray-800"
-                        required
-                      />
-                    </div>
-                    <div className="ml-3 text-sm">
-                      <label
-                        htmlFor="remember"
-                        className="text-gray-500 dark:text-gray-300"
-                      >
-                        Remember me
-                      </label>
-                    </div>
-                  </div>
+                  
                   <a
                     href="#sdnfjn"
                     className="text-sm font-medium text-primary-600 hover:underline dark:text-primary-500"
@@ -130,20 +150,24 @@ const Login: React.FC = () => {
                 </div>
                 <button
                   type="submit"
-                  className="w-full text-white bg-primary-600 hover:bg-primary-700 focus:ring-4 focus:outline-none focus:ring-primary-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800"
+                  className="w-full text-white bg-blue-600 hover:bg-primary-700 focus:ring-4 focus:outline-none focus:ring-primary-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800"
+                disabled={!isValid}
                 >
                   Sign in
                 </button>
-                <p className="text-sm font-light text-gray-500 dark:text-gray-400">
+             
+              </form>
+              <GoogleButton onClick={googleData}/>
+              <p className="text-sm font-light text-gray-500 dark:text-gray-400">
                   Don’t have an account yet?{" "}
-                  <a
-                    href="#ejn"
+                  <Link
+                    href="/register"
                     className="font-medium text-primary-600 hover:underline dark:text-primary-500"
                   >
                     Sign up
-                  </a>
+                  </Link>
                 </p>
-              </form>
+             
             </div>
           </div>
         </div>
